build(cloudformation): add vatertime/minecraft-spot-pricing cf file with some adjustment (need some experimentation)
Update the EFS to be One Zone. Add Tag in AutoScaling Group so that the Lambda function can find the EC2 instance and get the public IP Address via Discord command. Add the family of t4g and t3a instance type, starting with t4g.micro. Spot pricing is reduced to the minimum value to follow t4g.micro on-demand pricing. Refactor misleading parameter (MinecraftImageTag), changed to DockerImageTag instead to not confuse with the Minecraft Version.